finish authorization section

+++ src/components/features/ticket/actions/delete-ticket.ts
@@ -4,14 +4,28 @@ import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 
+import { isOwner } from "../../auth/utils/is-owner";
+import { getAuthOrRedirect } from "../queries/get-auth-or-redirect";
+
 export const deleteTicket = async (id: string) => {
-  await new Promise((resolve) => setTimeout(resolve, 2000));
+  const { user } = await getAuthOrRedirect();
 
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("Not authorized", "ERROR");
+    }
     await prisma.ticket.delete({
       where: {
         id,
